refactor: don't expose kafka helpers in public module
